+++ config/auth.js
@@ -1,4 +1,5 @@
 module.exports = {
   saltRounds: 10,
   saltSecret: process.env.AUTH_SALT || 'gdfg45667_%%^trterte',
+  isAuthWithVefificationFlow: true,
 };