+++ controllers/Auth.test.js
@@ -3,6 +3,8 @@ const request = require('supertest');
 const userEmail = 'testing@test.com';
 const userPassword = 'SuperNiceSecret123$';
 
+const userEmail2 = 'testing2@test.com';
+
 describe('auth', () => {
   describe('registration', () => {
     it('code NOT able to create user with wrong email', async () => {
@@ -28,7 +30,7 @@ describe('auth', () => {
           password: userPassword,
           nickName: 'test',
         });
-      expect(status).toBe(200);
+      expect(status).toBe(201);
       expect(body.success).toBe(true);
     });
 
@@ -68,7 +70,7 @@ describe('auth', () => {
       expect(status).toBe(400);
     });
 
-    it('can  login with normal creds and  verifyed email', async () => {
+    it('can login with normal creds and verifyed email', async () => {
       expect.assertions(3);
 
       const user = await global.server.app
@@ -89,7 +91,46 @@ describe('auth', () => {
       expect(body.success).toBe(true);
       expect(body.token).toBeDefined();
     });
+
+    it('can login with normal creds and  NOT verifyed email is option isAuthWithVefificationFlow is set', async () => {
+      expect.assertions(5);
+
+      const { status } = await request(global.server.app.httpServer.express)
+        .post('/auth/register')
+        .send({
+          email: userEmail2,
+          password: userPassword,
+        });
+
+      const { status: status2 } = await request(
+        global.server.app.httpServer.express,
+      )
+        .post('/auth/login')
+        .send({
+          email: userEmail2,
+          password: userPassword,
+        });
+
+      global.server.app.updateConfig('Auth', {
+        isAuthWithVefificationFlow: false,
+      });
+
+      const { status: status3, body } = await request(
+        global.server.app.httpServer.express,
+      )
+        .post('/auth/login')
+        .send({
+          email: userEmail2,
+          password: userPassword,
+        });
+      expect(status).toBe(201);
+      expect(status2).toBe(400);
+      expect(status3).toBe(200);
+      expect(body.success).toBe(true);
+      expect(body.token).toBeDefined();
+    });
   });
+
   describe('rate limiter', () => {
     it('we  should receive 429 on rate limit exceeded', async () => {
       expect.assertions(1);